Reorganize code base
Reorganize classes to be in relevant packages.
