+++ src/main/java/seedu/address/logic/parser/CreateTaskCommandParser.java
@@ -8,7 +8,7 @@ import java.util.List;
 import java.util.Set;
 import java.util.stream.Stream;
 
-import seedu.address.logic.commands.CreateTaskCommand;
+import seedu.address.logic.commands.task.CreateTaskCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.task.Task;
