+++ src/main/java/seedu/address/logic/parser/DeleteTagCommandParser.java
@@ -6,7 +6,7 @@ import static seedu.address.logic.parser.CliSyntax.PREFIX_TAG;
 
 import java.util.stream.Stream;
 
-import seedu.address.logic.commands.DeleteTagCommand;
+import seedu.address.logic.commands.tag.DeleteTagCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.tag.Tag;
