+++ src/main/java/seedu/address/logic/parser/DeleteTaskCommandParser.java
@@ -3,7 +3,7 @@ package seedu.address.logic.parser;
 import static seedu.address.logic.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 
 import seedu.address.commons.core.index.Index;
-import seedu.address.logic.commands.DeleteTaskCommand;
+import seedu.address.logic.commands.task.DeleteTaskCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
 
 /**
