+++ src/test/java/seedu/address/logic/commands/person/AddCommandTest.java
@@ -1,11 +1,11 @@
-package seedu.address.logic.commands;
+package seedu.address.logic.commands.person;
 
 import static java.util.Objects.requireNonNull;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
-import static seedu.address.logic.commands.CommandTestUtil.VALID_TAG_FLORIST;
-import static seedu.address.logic.commands.CommandTestUtil.VALID_TAG_PHOTOGRAPHER;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 import static seedu.address.testutil.Assert.assertThrows;
+import static seedu.address.testutil.TypicalPersons.ALICE;
 
 import java.nio.file.Path;
 import java.util.ArrayList;
@@ -17,6 +17,7 @@ import org.junit.jupiter.api.Test;
 import javafx.collections.ObservableList;
 import seedu.address.commons.core.GuiSettings;
 import seedu.address.logic.Messages;
+import seedu.address.logic.commands.CommandResult;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.AddressBook;
 import seedu.address.model.Model;
@@ -25,60 +26,67 @@ import seedu.address.model.ReadOnlyUserPrefs;
 import seedu.address.model.person.Person;
 import seedu.address.model.person.Vendor;
 import seedu.address.model.tag.Tag;
-import seedu.address.model.tag.TagName;
 import seedu.address.model.task.Task;
 import seedu.address.model.wedding.Wedding;
+import seedu.address.testutil.PersonBuilder;
 
-public class CreateTagCommandTest {
+public class AddCommandTest {
 
     @Test
-    public void constructor_nullTag_throwsNullPointerException() {
-        assertThrows(NullPointerException.class, () -> new CreateTagCommand(null));
+    public void constructor_nullPerson_throwsNullPointerException() {
+        assertThrows(NullPointerException.class, () -> new AddCommand(null));
     }
 
     @Test
-    public void execute_tagAcceptedByModel_addSuccessful() throws Exception {
-        ModelStubAcceptingTagAdded modelStub = new ModelStubAcceptingTagAdded();
-        Tag validTag = new Tag(new TagName(VALID_TAG_FLORIST));
+    public void execute_personAcceptedByModel_addSuccessful() throws Exception {
+        ModelStubAcceptingPersonAdded modelStub = new ModelStubAcceptingPersonAdded();
+        Person validPerson = new PersonBuilder().build();
 
-        CommandResult commandResult = new CreateTagCommand(validTag).execute(modelStub);
-        assertEquals(String.format(CreateTagCommand.MESSAGE_SUCCESS, Messages.format(validTag)),
+        CommandResult commandResult = new AddCommand(validPerson).execute(modelStub);
+
+        assertEquals(String.format(AddCommand.MESSAGE_SUCCESS, Messages.format(validPerson)),
                 commandResult.getFeedbackToUser());
-        assertEquals(Arrays.asList(validTag), modelStub.tagsAdded);
+        assertEquals(Arrays.asList(validPerson), modelStub.personsAdded);
     }
 
     @Test
-    public void execute_duplicateTag_throwsCommandException() {
-        Tag validTag = new Tag(new TagName(VALID_TAG_FLORIST));
-        CreateTagCommand createTagCommand = new CreateTagCommand(validTag);
-        ModelStub modelStub = new ModelStubWithTag(validTag);
+    public void execute_duplicatePerson_throwsCommandException() {
+        Person validPerson = new PersonBuilder().build();
+        AddCommand addCommand = new AddCommand(validPerson);
+        ModelStub modelStub = new ModelStubWithPerson(validPerson);
 
-        assertThrows(CommandException.class,
-                CreateTagCommand.MESSAGE_DUPLICATE_TAG, () -> createTagCommand.execute(modelStub));
+        assertThrows(CommandException.class, AddCommand.MESSAGE_DUPLICATE_PERSON, () -> addCommand.execute(modelStub));
     }
 
     @Test
     public void equals() {
-        Tag florist = new Tag(new TagName(VALID_TAG_FLORIST));
-        Tag photographer = new Tag(new TagName(VALID_TAG_PHOTOGRAPHER));
-        CreateTagCommand createFloristCommand = new CreateTagCommand(florist);
-        CreateTagCommand createPhotographerCommand = new CreateTagCommand(photographer);
+        Person alice = new PersonBuilder().withName("Alice").build();
+        Person bob = new PersonBuilder().withName("Bob").build();
+        AddCommand addAliceCommand = new AddCommand(alice);
+        AddCommand addBobCommand = new AddCommand(bob);
 
-        // same object -> returns ture
-        assertEquals(createFloristCommand, createFloristCommand);
+        // same object -> returns true
+        assertTrue(addAliceCommand.equals(addAliceCommand));
 
-        // same values -> returns false
-        CreateTagCommand createFloristCommandCopy = new CreateTagCommand(florist);
-        assertEquals(createFloristCommand, createFloristCommandCopy);
+        // same values -> returns true
+        AddCommand addAliceCommandCopy = new AddCommand(alice);
+        assertTrue(addAliceCommand.equals(addAliceCommandCopy));
 
         // different types -> returns false
-        assertFalse(createFloristCommand.equals(1));
+        assertFalse(addAliceCommand.equals(1));
 
-        // null -> return false
-        assertFalse(createFloristCommand.equals(null));
+        // null -> returns false
+        assertFalse(addAliceCommand.equals(null));
 
-        // different tag -> return false
-        assertFalse(createFloristCommand.equals(createPhotographerCommand));
+        // different person -> returns false
+        assertFalse(addAliceCommand.equals(addBobCommand));
+    }
+
+    @Test
+    public void toStringMethod() {
+        AddCommand addCommand = new AddCommand(ALICE);
+        String expected = AddCommand.class.getCanonicalName() + "{toAdd=" + ALICE + "}";
+        assertEquals(expected, addCommand.toString());
     }
 
     /**
@@ -166,17 +174,22 @@ public class CreateTagCommandTest {
         }
 
         @Override
-        public void updateFilteredPersonListByTag(Predicate<Tag> tagPredicate) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
         public boolean hasTag(Tag toAdd) {
             throw new AssertionError("This method should not be called.");
         }
 
         @Override
         public void addTag(Tag toAdd) {
+            throw new AssertionError("This method should not be called.");
+        }
+
+        @Override
+        public void updateFilteredTagList(Predicate<Tag> predicate) {
+            throw new AssertionError("This method should not be called.");
+        }
+
+        @Override
+        public ObservableList<Tag> getFilteredTagList() {
             throw new AssertionError("This method should not be called.");
         }
 
@@ -221,12 +234,12 @@ public class CreateTagCommandTest {
         }
 
         @Override
-        public void setWedding(Wedding target, Wedding editedWedding) {
+        public void deleteWedding(Wedding toDelete) {
             throw new AssertionError("This method should not be called.");
         }
 
         @Override
-        public void deleteWedding(Wedding tag) {
+        public void setWedding(Wedding target, Wedding editedWedding) {
             throw new AssertionError("This method should not be called.");
         }
 
@@ -241,22 +254,12 @@ public class CreateTagCommandTest {
         }
 
         @Override
-        public void updateFilteredPersonListByWedding(Predicate<Wedding> tag) {
+        public void deleteTag(Tag toDelete) {
             throw new AssertionError("This method should not be called.");
         }
 
         @Override
-        public void updateFilteredTagList(Predicate<Tag> predicate) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ObservableList<Tag> getFilteredTagList() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void deleteTag(Tag tag) {
+        public void updateFilteredPersonListByTag(Predicate<Tag> tag) {
             throw new AssertionError("This method should not be called.");
         }
 
@@ -274,42 +277,48 @@ public class CreateTagCommandTest {
         public void unassignVendor(Person person) {
             throw new AssertionError("This method should not be called.");
         }
+
+        @Override
+        public void updateFilteredPersonListByWedding(Predicate<Wedding> tag) {
+            throw new AssertionError("This method should not be called.");
+        }
+
     }
 
     /**
-     * A Model stub that contains a single tag.
+     * A Model stub that contains a single person.
      */
-    private class ModelStubWithTag extends CreateTagCommandTest.ModelStub {
-        private final Tag tag;
+    private class ModelStubWithPerson extends ModelStub {
+        private final Person person;
 
-        ModelStubWithTag(Tag tag) {
-            requireNonNull(tag);
-            this.tag = tag;
+        ModelStubWithPerson(Person person) {
+            requireNonNull(person);
+            this.person = person;
         }
 
         @Override
-        public boolean hasTag(Tag tag) {
-            requireNonNull(tag);
-            return this.tag.isSameTag(tag);
+        public boolean hasPerson(Person person) {
+            requireNonNull(person);
+            return this.person.isSamePerson(person);
         }
     }
 
     /**
-     * A Model stub that always accept the tag being added.
+     * A Model stub that always accept the person being added.
      */
-    private class ModelStubAcceptingTagAdded extends CreateTagCommandTest.ModelStub {
-        final ArrayList<Tag> tagsAdded = new ArrayList<>();
+    private class ModelStubAcceptingPersonAdded extends ModelStub {
+        final ArrayList<Person> personsAdded = new ArrayList<>();
 
         @Override
-        public boolean hasTag(Tag tag) {
-            requireNonNull(tag);
-            return tagsAdded.stream().anyMatch(tag::isSameTag);
+        public boolean hasPerson(Person person) {
+            requireNonNull(person);
+            return personsAdded.stream().anyMatch(person::isSamePerson);
         }
 
         @Override
-        public void addTag(Tag tag) {
-            requireNonNull(tag);
-            tagsAdded.add(tag);
+        public void addPerson(Person person) {
+            requireNonNull(person);
+            personsAdded.add(person);
         }
 
         @Override
@@ -317,4 +326,5 @@ public class CreateTagCommandTest {
             return new AddressBook();
         }
     }
+
 }
