+++ src/test/java/seedu/address/logic/commands/task/CreateTaskCommandTest.java
@@ -1,4 +1,4 @@
-package seedu.address.logic.commands;
+package seedu.address.logic.commands.task;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
@@ -12,6 +12,7 @@ import java.util.HashSet;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
+import seedu.address.logic.commands.CommandResult;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
@@ -77,7 +78,7 @@ public class CreateTaskCommandTest {
 
         CreateTaskCommand command = new CreateTaskCommand(tasksToAdd);
 
-        String expectedString = "seedu.address.logic.commands.CreateTaskCommand{taskToAdd=[[T][ ] Buy groceries]}";
+        String expectedString = "seedu.address.logic.commands.task.CreateTaskCommand{taskToAdd=[[T][ ] Buy groceries]}";
         assertEquals(expectedString, command.toString());
     }
