+++ src/test/java/seedu/address/logic/parser/AddCommandParserTest.java
@@ -52,7 +52,7 @@ import static seedu.address.testutil.TypicalPersons.ERIC;
 import org.junit.jupiter.api.Test;
 
 import seedu.address.logic.Messages;
-import seedu.address.logic.commands.AddCommand;
+import seedu.address.logic.commands.person.AddCommand;
 import seedu.address.model.person.Email;
 import seedu.address.model.person.Name;
 import seedu.address.model.person.Person;
