+++ src/test/java/seedu/address/logic/parser/CreateTagCommandParserTest.java
@@ -11,7 +11,7 @@ import static seedu.address.logic.parser.CommandParserTestUtil.assertParseSucces
 import org.junit.jupiter.api.Test;
 
 import seedu.address.logic.Messages;
-import seedu.address.logic.commands.CreateTagCommand;
+import seedu.address.logic.commands.tag.CreateTagCommand;
 import seedu.address.model.tag.Tag;
 import seedu.address.model.tag.TagName;
