+++ src/test/java/seedu/address/logic/parser/CreateTaskCommandParserTest.java
@@ -14,7 +14,7 @@ import java.util.Set;
 import org.junit.jupiter.api.Test;
 
 import seedu.address.logic.Messages;
-import seedu.address.logic.commands.CreateTaskCommand;
+import seedu.address.logic.commands.task.CreateTaskCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.task.Task;
 import seedu.address.testutil.TypicalTasks;
