+++ src/test/java/seedu/address/logic/parser/UntagCommandParserTest.java
@@ -11,7 +11,7 @@ import java.util.HashSet;
 import org.junit.jupiter.api.Test;
 
 import seedu.address.commons.core.index.Index;
-import seedu.address.logic.commands.UntagCommand;
+import seedu.address.logic.commands.tag.UntagCommand;
 import seedu.address.model.tag.Tag;
 import seedu.address.model.tag.TagName;
